V2. G001. Redone refresh to newer software versions  (#42)
## Update of README and chapter G001

This update corrects the contents of the READMe and the chapter G001 so they talk about a real hardware setup rather than a virtualized one (which was discarded due to issues with the networking setup). Also, it changes mentions of "guide series" with just "guide", and speaks of "chapters" rather than "guides".