V2. G001. Redone (again) refresh to newer software versions (#43)
## Update of README and chapter G001

This update changes again the contents of the README and the chapter G001 to ensure they talk about a real hardware setup rather than a virtualized one (which was discarded due to issues with the networking setup). In particular, adds a mention to the UEFI BIOS and a different UPS unit. Also, changes in the README sections, corrections in contents and fixes reference links.