V2. G002. Redone (again) refresh to newer software versions (#44)
## Yet another update of chapter G002

This update changes the Proxmox VE installation procedure so it is about Proxmox VE 9.0. This implied replacing the Rufus procedure with the mention to the Ventoy tool, correcting all the texts and refreshing all snapshots. Also adds a subsection explaining how to solve the possible error of the EFI bootloader.