V2. G017. Redone refresh to newer software versions (#60)
## Another update of chapter G017

This reviews the content of the chapter G017 to update the snapshots so they correspond to Proxmox VE 9.0 plus makes some corrections in the text. Also fixes a mistake in the G014 chapter because it was related to networking (the name of the ethernet NIC was incorrect like in the G017 chapter).