V2. G020. Refresh to newer software versions (#63)
## Update of chapter G020

Refreshes the snapshots and procedures to fit Proxmox VE 9.0 and Debian 13. In particular, in the installation process of Debian in the VM, the partition schema chosen now is the one meant `for servers`. This sets the `/var` and `/srv` directories, and the swap storage as separated partitions managed by LVM.