V2. G025 and G026. Fixes issue with external IP and K3s server node TLC cert (#72)
## Correction of chapters G025 and G026

Corrects an issue with the external IP of the K3s server node. It is necessary for it to be specified also in the tls-san option of the server node configuration. Otherwise, is rejected by the autogenerated certificate when trying to connect with the `kubectl` client through that external IP. This has implied doing fixes both in the chapter G025 and G026.

Also fixes the use of the "guide" term and other mistakes.