V2. G029. Simplifies cert-manager deployment (#78)
## Correction of chapter G029

In a previous version, I had customized the default namespace where cert-manage would look up for certificates in the cluster. Later, I discovered that this was not really necessary when using cluster issuers so I decided to simplify the cert-manager setup and leave it with its original default configuration. Also, I've taken the chance to standardize better the names used in the certificates and use the `-tls` suffix rather than `-crt`. All these considerations have implied the following changes:

- Simplification of the Kustomize deployment of cert-manager by removing any setup related to change the default namespace for certificates.
- Renaming of the CA certificates to use the `-tls` suffix, which also implied renaming the corresponding YAML files.
- Removal of an entire section about cert-manager performance in deployment, mainly because I didn't had the issue this time.
- Bumping up of cert-manager version used to v1.19.0.
- Several other corrections in the chapter G029, mainly in the procedures's steps and in the shell snippets.